Stop openssh from creating ~/.rnd
openssh creates an entropy file at ${HOME}/.rnd which, in the docker
container, is owned by root. When this file is in a user-owned
directory (or home) it can stomp on the user-owned file by the same
name.

By setting the RANDFILE environment variable, we can force it to use a
safe filename.