Document how to map the needed files instead of $HOME
Add documentation to the README.md for binding just the files we need
for K2 instead of the entire home directory. This prevents the K2
processes from writing potentially overwriting files in the user's home
directory when running as a container.

Fixes #181 and is associated with #97

Similar changes to k2cli should be done,
https://github.com/samsung-cnct/k2cli/issues/51