prevent kraken from removing some kubernetes created resources
kraken creates the vpc route table and the vpc security group , however
kubernetes will make modifications to both assets as the cluster runs.
we should not let kraken modify either of these resources going forward.

there is no way to specify custom routes in the config file so we are fine there
we need to modify the documentation on the security group to specify that
the ingress and egress rules for the security will only be applied at cluster
creation.